refactor(web): 이미지 재시도 상태 머신을 useImageRetry 훅으로 추출
- PhotoThumbnail·PhotoProgressiveImage에 중복된 재시도 로직을 한곳으로 모음
- 재시도 정책(자동 1회·수동 재시도 순서)을 훅에서만 관리해 갈라짐 방지
- 두 컴포넌트에는 각자의 실패 UI(아이콘 / 문구+버튼)만 남김
